refactor(ui): split ReviewHeader into HeaderBrand, HeaderSource, HeaderProgress, HeaderActions
